Bump version to 0.8.1.1
When trying to do `cabal install cmv`, I found ghc trying and failing to build `PrimitiveArray-0.8.0.1`. This is surprising because the current `BiobaseInfernal.cabal` does not have 0.8.0.1. It turns out  `BiobaseInfernal.cabal` was changed after 0.8.1.0 was released without bumping the version number. This presumably causes cabal to still have the old version file in their database? I hope this fixes the issue.